Encapsulation and Unit Tests for members
