+++ members/testdata/members_test.csv
@@ -1,0 +1,15 @@
+Name,Street,City,State,Zip,County,Phone,Email,Last_One_On_One,Issues,Due_Date_Pay,Active
+John Doe,345 Main St,Thomasville,NC,49583,Shoresville,555-8439,john@example.com,04/13/25,"housing, education",08/06/20,TRUE
+Jane Doe,947 Green St,Shelby,IL,50249,Tinker,555-8564,jane@example.com,08/20/24,"criminal justice, housing",06/25/21,FALSE
+Robert Marks,693 Tyler Dr.,Alamance,OH,29423,Brown,543-8543,robert@example.com,06/14/12,"climate change, education",06/20/20,FALSE
+Graham Norton,255 Parker Cir,Todd,SC,15482,Toddville,225-8457,graham@example.com,04/01/25,"climate change, economy",04/03/19,TRUE
+Stephen King,666 Dark Dr.,Salem,MA,82734,Crow,555-8264,stephen@example.com,04/01/25,criminal justice,04/01/25,TRUE
+Harry Potter,548 Hedwig Dr.,Potts,UK,45873,Magical,845-3654,harry@example.com,03/25/24,"education, healthcare",08/20/24,FALSE
+Crash Bandicoot,132 Faster Ct.,Speed,CA,90210,Los Angeles,555-4587,crash@example.com,04/15/25,"mental health, housing",12/25/20,TRUE
+Gene Simmons,547 Grammer Dr.,Topsy,WI,54578,Greene,547-8523,gene@example.com,10/18/13,"housing, criminal justice",03/23/02,FALSE
+Tyler Perry,654 Dream St.,Atlanta,GA,54785,Fulton,543-8562,tyler@example.com,03/21/25,"education, healthcare",05/25/23,TRUE
+Bob Hope,852 Pope St.,Mitchell,RI,45623,Banks,458-6592,bob@example.com,11/20/24,"healthcare, economy",02/22/25,FALSE
+Billy Bob Thorton,304 Easy St.,Burlington,TX,65842,Harris,523-6652,billy@example.com,09/01/22,"economy, housing",10/12/00,TRUE
+Marianne Webster,400 Rainbow Rd.,Tombs,TN,65478,Shiner,542-8856,marianne@example.com,08/05/24,"climate change, criminal justice",05/30/24,FALSE
+Jeremy Timber,654 Gabby Ln,Ross,NC,25523,Caswell,652-3323,jeremy@example.com,04/15/25,"education, housing",04/01/25,TRUE
+James McAvoy,447 Black Dr.,Greensboro,IN,44758,Brooks,445-8859,james@example.com,03/29/25,"healthcare, economy",03/23/24,TRUE
